+++ Connector/IdentityService/IdentityService.php
@@ -5,27 +5,32 @@ namespace SystemConnector\IdentityService;
 use Assert\Assertion;
 use Ramsey\Uuid\Uuid;
 use SystemConnector\IdentityService\Exception\NotFoundException;
-use SystemConnector\IdentityService\Storage\IdentityStorageInterface;
+use SystemConnector\IdentityService\Storage\IdentityServiceStorageInterface;
 use SystemConnector\ValidatorService\ValidatorServiceInterface;
 use SystemConnector\ValueObject\Identity\Identity;
+use Traversable;
 
 class IdentityService implements IdentityServiceInterface
 {
     /**
-     * @var IdentityStorageInterface
+     * @var IdentityServiceStorageInterface[]|Traversable
      */
-    private $storage;
+    private $storages;
 
     /**
      * @var ValidatorServiceInterface
      */
     private $validator;
 
+    /**
+     * @param IdentityServiceStorageInterface[]|Traversable $storage
+     * @param ValidatorServiceInterface                     $validator
+     */
     public function __construct(
-        IdentityStorageInterface $storage,
+        Traversable $storage,
         ValidatorServiceInterface $validator
     ) {
-        $this->storage = $storage;
+        $this->storages = iterator_to_array($storage);
         $this->validator = $validator;
     }
 
@@ -48,8 +53,12 @@ class IdentityService implements IdentityServiceInterface
         ]);
 
         if (null === $identity) {
-            throw new NotFoundException(sprintf('Could not find identity for %s with identifier %s in %s.', $objectType,
-                $adapterIdentifier, $adapterName));
+            throw new NotFoundException(printf(
+                'Could not find identity for %s with identifier %s in %s.',
+                $objectType,
+                $adapterIdentifier,
+                $adapterName
+            ));
         }
 
         $this->validator->validate($identity);
@@ -78,7 +87,7 @@ class IdentityService implements IdentityServiceInterface
         if (null === $identity) {
             $objectIdentifier = Uuid::uuid4()->toString();
 
-            $identity = $this->create(
+            $identity = $this->insert(
                 $objectIdentifier,
                 $objectType,
                 (string) $adapterIdentifier,
@@ -98,31 +107,37 @@ class IdentityService implements IdentityServiceInterface
     {
         Assertion::isArray($criteria);
 
-        $identity = $this->storage->findOneBy($criteria);
-        $this->validator->validate($identity);
+        $storage = reset($this->storages);
+        $identity = $storage->findOneBy($criteria);
 
-        return $identity;
+        if ($identity !== null) {
+            $this->validator->validate($identity);
+
+            return $identity;
+        }
+
+        return null;
     }
 
     /**
      * {@inheritdoc}
      */
-    public function create($objectIdentifier, $objectType, $adapterIdentifier, $adapterName)
+    public function insert($objectIdentifier, $objectType, $adapterIdentifier, $adapterName)
     {
-        $params = compact(
-            'objectIdentifier',
-            'objectType',
-            'adapterIdentifier',
-            'adapterName'
-        );
-
         /**
          * @var Identity $identity
          */
-        $identity = Identity::fromArray($params);
+        $identity = Identity::fromArray([
+            'objectIdentifier' => $objectIdentifier,
+            'objectType' => $objectType,
+            'adapterIdentifier' => $adapterIdentifier,
+            'adapterName' => $adapterName,
+        ]);
 
         $this->validator->validate($identity);
-        $this->storage->persist($identity);
+
+        $storage = reset($this->storages);
+        $storage->insert($identity);
 
         return $identity;
     }
@@ -134,7 +149,8 @@ class IdentityService implements IdentityServiceInterface
     {
         Assertion::isArray($criteria);
 
-        $identities = $this->storage->findBy($criteria);
+        $storage = reset($this->storages);
+        $identities = $storage->findBy($criteria);
 
         array_walk($identities, function (Identity $identity) {
             $this->validator->validate($identity);
@@ -149,7 +165,28 @@ class IdentityService implements IdentityServiceInterface
     public function update(Identity $identity, array $params = [])
     {
         $this->validator->validate($identity);
-        $this->storage->update($identity, $params);
+
+        $newIdentity = clone $identity;
+
+        if (!empty($params['objectIdentifier'])) {
+            $newIdentity->setObjectIdentifier($params['objectIdentifier']);
+        }
+        if (!empty($params['objectType'])) {
+            $newIdentity->setAdapterName($params['objectType']);
+        }
+        if (!empty($params['adapterIdentifier'])) {
+            $newIdentity->setAdapterIdentifier($params['adapterIdentifier']);
+        }
+        if (!empty($params['adapterName'])) {
+            $newIdentity->setObjectType($params['adapterName']);
+        }
+
+        $this->validator->validate($newIdentity);
+
+        $storage = reset($this->storages);
+        $storage->update($identity, $params);
+
+        return $newIdentity;
     }
 
     /**
@@ -158,7 +195,9 @@ class IdentityService implements IdentityServiceInterface
     public function remove(Identity $identity)
     {
         $this->validator->validate($identity);
-        $this->storage->remove($identity);
+
+        $storage = reset($this->storages);
+        $storage->remove($identity);
     }
 
     /**