+++ Installer/Model/Config.php
@@ -1,20 +1,22 @@
 <?php
 
-namespace SystemConnector\ConfigService\Model;
+namespace PlentyConnector\Installer\Model;
 
 use Doctrine\ORM\Mapping as ORM;
 
 /**
- * @ORM\Entity(repositoryClass="ConfigRepository")
- * @ORM\Table(name="plenty_config")
+ * @ORM\Entity
+ * @ORM\Table(name="plenty_config", indexes={
+ *     @ORM\Index(name="name_idx", columns={"name"})
+ * })
  */
 class Config
 {
     /**
      * @var int
      *
-     * @ORM\Column(name="id", type="integer", nullable=false)
      * @ORM\Id
+     * @ORM\Column(name="id", type="integer", nullable=false)
      * @ORM\GeneratedValue(strategy="IDENTITY")
      */
     private $id;